Allows to check for keys associated to NaN
The following should not raise an assertion exception

    assert.propertyVal({camus: NaN}, 'camus', NaN)

But, since [NaN is not equal to itself as it ought to
be](http://stackoverflow.com/questions/1565<t_co>4/what-is-the-rationale-for-all-comparisons-returning-false-for-ieee754-nan-values)
the previous assertion wasn't behaving as expected.

Now, I'm not sure that the weird semantics of `NaN` should allow for
such a thing, but I'm inclined to believe that, if anyone wants to
assert that they're expecting `NaN` to be associated to a value, its
startling nature shouldn't get in the way.

 