fix: simplify chai module export
There's an issue around chai using the `exports` keyword in a dynamic way that means the main module cannot be used in a way with module systems that statically infer what the exports are. 

This change simply refactors the export code to ensure `chai` is a single module that gets exported once, and doesn't refer to `module.exports` or `exports` inside of any function calls.