docs: Add type definitions
This commit adds type definitions for Chai. These type definitions are
based on the types from DefinitelyTyped. They enable Typescript users
to use the library directly and serve as documentation.

The main changes from the DT version are:
- Use external module style
- Use `AssertionError` types from `assertion-error`
- Fix global modification of objects with `should`

The types can be improved in further PRs.

This commit also adds tests for the type definitions. It tries to
compile (in noEmit mode) the files in `test/typings` using the type
definitions. This adds `typescript` as a new dev dependency.

For maintenance and semver, the types should be treated as
documentation. They should be kept up-to-date with the code, especially
when breaking changes occur (wrong documentation is worse than
incomplete documentation), see discussion in #1092.
[Help for type definitions][ts-declarations]

- Closes chaijs/chai#650 (original issue)
- See chaijs/chai#822 (previous PR with merge conflicts)
- Closes chaijs/chai#1092 (superseded by this PR)

[ts-declarations]: https://www.typescriptlang.org/docs/handbook/declaration-files/introduction.html