feat: rework into ES modules
This moves all of the sources to be ES modules rather than CommonJS.

In order to produce the entrypoints, we use esbuild as a bundler (and as
a transpiler in a way).

Due to the fact that some dependencies are written in CommonJS, we
actually use esbuild to create _two_ bundles: a CommonJS bundle, and an
ES module bundle.

Otherwise, someone importing the raw source would inevitably end up
trying to import a CommonJS module somewhere down the tree, which would
not work in browsers natively.